run all web socket tests in Node
