--- src/socket/debug/test/_/start-server.ts
@@ -1,31 +1,0 @@
-import { createServer } from 'http';
-
-import { MuWebSocketServer } from '../../../web/server';
-import { MuDebugServer } from '../../index';
-import { MuServer } from '../../../../server';
-import { protocolSchema } from './schema';
-
-const PING_OPCODE = 0x9;
-const PONG_OPCODE = 0xA;
-
-const server = createServer();
-const socketServer = new MuWebSocketServer({ server });
-const debugSocketServer = new MuDebugServer({ socketServer });
-const muServer = new MuServer(debugSocketServer);
-
-muServer.protocol(protocolSchema).configure({
-    message: {
-        ping: (client, data, unreliable) => {
-            const response = data === PING_OPCODE ? PONG_OPCODE : data;
-            client.message.pong(response, !!unreliable);
-        },
-    },
-});
-
-const port = process.argv[2];
-server.listen(port);
-muServer.start({
-    ready: () => {
-        console.log(`server listening on port ${server.address().port}...`);
-    },
-});
